+++ packages/layout/lib/src/modal/components/modal-overlay/modal-overlay.component.spec.ts
@@ -1,11 +1,17 @@
+import { Component, DebugElement } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { DebugElement, Component } from '@angular/core';
-import { BrowserDynamicTestingModule } from '@angular/platform-browser-dynamic/testing';
-
+import { By } from '@angular/platform-browser';
 import { ModalOverlayComponent } from './modal-overlay.component';
 
 const getByCSSQuery = (query, elm, all?) => {
 	return all ? elm.querySelectorAll(query) : elm.querySelector(query);
+};
+
+const mousedown = (element) => {
+	element.dispatchEvent(new MouseEvent('mousedown', { bubbles: true }));
+};
+const mouseup = (element) => {
+	element.dispatchEvent(new MouseEvent('mouseup', { bubbles: true }));
 };
 
 @Component({
@@ -19,7 +25,7 @@ const getByCSSQuery = (query, elm, all?) => {
         </aui-modal-overlay>
     `,
 })
-class TestComponent {}
+class TestComponent { }
 
 const modalMock = {
 	instance: {
@@ -32,6 +38,7 @@ describe('Modal - ModalOverlayComponent', () => {
 	let fixture: ComponentFixture<TestComponent>;
 	let de: DebugElement;
 	let overlay: ModalOverlayComponent;
+	let outsideModalElement: DebugElement;
 
 	// async beforeEach
 	beforeEach(async(() => {
@@ -41,7 +48,7 @@ describe('Modal - ModalOverlayComponent', () => {
 				TestComponent,
 			],
 		})
-		.compileComponents();
+			.compileComponents();
 	}));
 
 	beforeEach(() => {
@@ -58,6 +65,7 @@ describe('Modal - ModalOverlayComponent', () => {
 		(overlay as any).modal = modalMock;
 
 		spyOn(modalMock.instance, 'closeModal');
+		outsideModalElement = de.query(By.css('#outsideModal'));
 	});
 
 	it('should set the overlay classes on the host element', () => {
@@ -65,9 +73,12 @@ describe('Modal - ModalOverlayComponent', () => {
 	});
 
 	it('should do nothing if the modal inside the overlay is clicked', () => {
-		const insideModal = getByCSSQuery('#insideModal', de.nativeElement);
 
-		insideModal.click();
+		const insideModal = getByCSSQuery('#insideModal', de.nativeElement);
+		const modal = getByCSSQuery('.m-modal', de.nativeElement);
+
+		mousedown(insideModal);
+		mouseup(modal);
 
 		fixture.detectChanges();
 
@@ -77,7 +88,8 @@ describe('Modal - ModalOverlayComponent', () => {
 	it('should close the modal if the overlay is clicked', () => {
 		const overlayEl = getByCSSQuery('.m-overlay', de.nativeElement);
 
-		overlayEl.click();
+		mousedown(overlayEl);
+		mouseup(overlayEl);
 
 		fixture.detectChanges();
 
@@ -87,10 +99,24 @@ describe('Modal - ModalOverlayComponent', () => {
 	it('should close the modal if an element inside the overlay, but outside the modal is clicked', () => {
 		const outsideModal = getByCSSQuery('#outsideModal', de.nativeElement);
 
-		outsideModal.click();
+		mousedown(outsideModal);
+		mouseup(outsideModal);
 
 		fixture.detectChanges();
 
 		expect(modalMock.instance.closeModal).toHaveBeenCalled();
 	});
+
+	it('should do nothing if the click starts inside the modal but ends outside the modal', () => {
+		const insideModal = getByCSSQuery('#insideModal', de.nativeElement);
+		const outsideModal = getByCSSQuery('#outsideModal', de.nativeElement);
+
+		mousedown(insideModal);
+		mouseup(outsideModal);
+
+		fixture.detectChanges();
+
+		expect(modalMock.instance.closeModal).not.toHaveBeenCalled();
+	});
+
 });